Add a step to allow for "slowly" entering text
With a tiny (1ms) delay between keystrokes.

This is to allow for testing some (poorly-written) websites that fail
when text is entered very quickly.

This commit includes a test for the new functionality, (with some
javascript code that will behave differently depending on if text is
entered quickly or relatively slowly).