Add a retry_on_timeout decorator and use it on the lowest-level visit step
We have seen cases where the PhantomJS webdriver appears to hang
indefinitely. We're not sure what's causing these, but other people
seem to be hitting them as well, see:

	https://github.com/ariya/phantomjs/issues/11526

We have been able to gget selenium to raise a TimeoutException by calling:

	browser.set_page_load_timeout(20)  # in seconds

For a complete solution, we need to respond to the TimeoutException by
retrying the page load. This commit adds that try/except block by
means of a new decorator named @retry_on_timeout, (which, other than
which exception is caught, and not doing any sleeping, is _very_
similar to the existing @wait_for decorator).